Refactor costs calculation
