+++ src/Controller/Api/EnergyTariffController.php
@@ -29,8 +29,9 @@ use App\Enums\ChannelFunction;
 use App\Enums\EnergyPriceComponent;
 use App\Enums\EnergyPriceUnit;
 use App\Model\ApiVersions;
-use App\Utils\DatabaseUtils;
-use App\Utils\DateUtils;
+use App\Model\MeasurementLogs\EnergyCostLogHydrator;
+use App\Model\MeasurementLogs\EnergyCostRowFetcher;
+use App\Model\MeasurementLogs\EnergyCostSummaryBuilder;
 use Assert\Assertion;
 use Doctrine\ORM\EntityManagerInterface;
 use FOS\RestBundle\Controller\Annotations as Rest;
@@ -40,9 +41,12 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class EnergyTariffController extends RestController {
-    private const RECORD_LIMIT_PER_REQUEST = 10000;
-
-    public function __construct(private readonly EntityManagerInterface $measurementLogsEntityManager) {
+    public function __construct(
+        private readonly EntityManagerInterface $measurementLogsEntityManager,
+        private readonly EnergyCostRowFetcher $energyCostRowFetcher,
+        private readonly EnergyCostLogHydrator $energyCostLogHydrator,
+        private readonly EnergyCostSummaryBuilder $energyCostSummaryBuilder,
+    ) {
     }
 
     /**
@@ -186,11 +190,11 @@ class EnergyTariffController extends RestController {
             (int)$request->query->get('afterTimestamp'),
             (int)$request->query->get('beforeTimestamp'),
             $request->query->get('order') !== 'ASC',
-            (int)$request->query->get('limit', self::RECORD_LIMIT_PER_REQUEST),
+            (int)$request->query->get('limit', $this->energyCostRowFetcher->getRecordLimitPerRequest()),
             (int)$request->query->get('offset', 0)
         );
 
-        return $this->view($this->hydrateCostLogs($rows));
+        return $this->view($this->energyCostLogHydrator->hydrateLogs($rows));
     }
 
     /**
@@ -202,9 +206,15 @@ class EnergyTariffController extends RestController {
         $this->ensureElectricityMeterChannel($channel);
         $afterTimestamp = (int)$request->query->get('afterTimestamp');
         $beforeTimestamp = (int)$request->query->get('beforeTimestamp');
-        $rows = $this->fetchAllElectricityCostRows($channel->getId(), $afterTimestamp, $beforeTimestamp);
 
-        return $this->view(array_values($this->buildEnergyCostSummaries($channel->getId(), $rows, $afterTimestamp, $beforeTimestamp)));
+        return $this->view(
+            $this->energyCostSummaryBuilder->buildSummaries(
+                $channel->getId(),
+                $afterTimestamp,
+                $beforeTimestamp,
+                $this->energyCostRowFetcher,
+            )
+        );
     }
 
     private function ensureElectricityMeterChannel(IODeviceChannel $channel): void {
@@ -219,562 +229,7 @@ class EnergyTariffController extends RestController {
         int $limit,
         int $offset
     ): array {
-        $entityManager = $this->getMeasurementLogsEntityManager();
-        $platform = DatabaseUtils::getPlatform($entityManager);
-        $order = $orderDesc ? 'DESC' : 'ASC';
-        $slotStartExpr = $platform === DatabaseUtils::PSQL
-            ? "b.date - INTERVAL '15 minutes'"
-            : 'DATE_SUB(b.date, INTERVAL 15 MINUTE)';
-        $dateTsExpr = $platform === DatabaseUtils::PSQL ? 'EXTRACT(EPOCH FROM b.date)::INTEGER' : 'UNIX_TIMESTAMP(b.date)';
-        $slotStartTsExpr = $platform === DatabaseUtils::PSQL ? "EXTRACT(EPOCH FROM ($slotStartExpr))::INTEGER" : "UNIX_TIMESTAMP($slotStartExpr)";
-
-        $where = 'WHERE d.channel_id = :channelId ';
-        if ($afterTimestamp > 0) {
-            $where .= 'AND d.date > :afterDate ';
-        }
-        if ($beforeTimestamp > 0) {
-            $where .= 'AND d.date < :beforeDate ';
-        }
-
-        $sql = "SELECT $dateTsExpr date_timestamp,
-                $slotStartTsExpr slot_start_timestamp,
-                b.date,
-                b.phase1_fae,
-                b.phase2_fae,
-                b.phase3_fae,
-                pa.profile_id,
-                tp.tariff_id,
-                rz.zone_code,
-                pp.id price_period_id,
-                ppi.component_code,
-                ppi.amount,
-                ppi.unit,
-                pp.currency,
-                pp.billing_period_length,
-                pp.billing_period_unit,
-                ((COALESCE(b.phase1_fae, 0) + COALESCE(b.phase2_fae, 0) + COALESCE(b.phase3_fae, 0)) / 1000.0) total_kwh,
-                (COALESCE(b.phase1_fae, 0) / 1000.0) phase1_kwh,
-                (COALESCE(b.phase2_fae, 0) / 1000.0) phase2_kwh,
-                (COALESCE(b.phase3_fae, 0) / 1000.0) phase3_kwh
-            FROM (
-                SELECT d.channel_id, d.date, d.phase1_fae, d.phase2_fae, d.phase3_fae
-                FROM supla_em_delta_log d
-                $where
-                ORDER BY d.date $order
-                LIMIT :limit OFFSET :offset
-            ) b
-            LEFT JOIN supla_energy_tariff_profile_assignment pa
-                ON pa.channel_id = b.channel_id
-            LEFT JOIN supla_energy_tariff_profile_tariff_period tp
-                ON tp.profile_id = pa.profile_id
-                AND (tp.valid_from IS NULL OR $slotStartExpr >= tp.valid_from)
-                AND (tp.valid_to IS NULL OR $slotStartExpr < tp.valid_to)
-            LEFT JOIN supla_energy_tariff_resolved_zone rz
-                ON rz.tariff_id = tp.tariff_id
-                AND $slotStartExpr >= rz.period_start
-                AND $slotStartExpr < rz.period_end
-            LEFT JOIN supla_energy_tariff_profile_price_period pp
-                ON pp.tariff_period_id = tp.id
-                AND (pp.valid_from IS NULL OR $slotStartExpr >= pp.valid_from)
-                AND (pp.valid_to IS NULL OR $slotStartExpr < pp.valid_to)
-            LEFT JOIN supla_energy_tariff_profile_price_item ppi
-                ON ppi.price_period_id = pp.id
-                AND ppi.unit = 'kWh'
-                AND (ppi.zone_code = rz.zone_code OR ppi.zone_code IS NULL)
-            ORDER BY b.date $order, ppi.component_code ASC";
-
-        $stmt = $entityManager->getConnection()->prepare($sql);
-        $stmt->bindValue('channelId', $channelId, 'integer');
-        if ($afterTimestamp > 0) {
-            $stmt->bindValue('afterDate', DateUtils::timestampToMysqlUtc($afterTimestamp), 'string');
-        }
-        if ($beforeTimestamp > 0) {
-            $stmt->bindValue('beforeDate', DateUtils::timestampToMysqlUtc($beforeTimestamp), 'string');
-        }
-        $stmt->bindValue('limit', min(max($limit, 1), self::RECORD_LIMIT_PER_REQUEST), 'integer');
-        $stmt->bindValue('offset', max($offset, 0), 'integer');
-
-        return $stmt->executeQuery()->fetchAllAssociative();
-    }
-
-    private function fetchAllElectricityCostRows(int $channelId, int $afterTimestamp, int $beforeTimestamp): array {
-        $offset = 0;
-        $rows = [];
-        do {
-            $batch = $this->fetchElectricityCostRows(
-                $channelId,
-                $afterTimestamp,
-                $beforeTimestamp,
-                false,
-                self::RECORD_LIMIT_PER_REQUEST,
-                $offset
-            );
-            $rows = array_merge($rows, $batch);
-            $offset += self::RECORD_LIMIT_PER_REQUEST;
-        } while (count($batch) === self::RECORD_LIMIT_PER_REQUEST);
-
-        return $rows;
-    }
-
-    private function hydrateCostLogs(array $rows): array {
-        $logs = [];
-        foreach ($rows as $row) {
-            $key = (string)$row['date_timestamp'];
-            if (!isset($logs[$key])) {
-                $phase1 = (int)$row['phase1_fae'];
-                $phase2 = (int)$row['phase2_fae'];
-                $phase3 = (int)$row['phase3_fae'];
-                $logs[$key] = [
-                    'dateTimestamp' => (int)$row['date_timestamp'],
-                    'slotStartTimestamp' => (int)$row['slot_start_timestamp'],
-                    'profileId' => $row['profile_id'] !== null ? (int)$row['profile_id'] : null,
-                    'tariffId' => $row['tariff_id'] !== null ? (int)$row['tariff_id'] : null,
-                    'zoneCode' => $row['zone_code'],
-                    'pricePeriodId' => $row['price_period_id'] !== null ? (int)$row['price_period_id'] : null,
-                    'usage' => [
-                        'phase1Fae' => $phase1,
-                        'phase2Fae' => $phase2,
-                        'phase3Fae' => $phase3,
-                        'totalFae' => $phase1 + $phase2 + $phase3,
-                        'totalKwh' => round(((float)$row['total_kwh']), 6),
-                    ],
-                    'costs' => null,
-                ];
-            }
-
-            if (!$row['component_code']) {
-                continue;
-            }
-
-            if ($logs[$key]['costs'] === null) {
-                $logs[$key]['costs'] = [
-                    'currency' => $row['currency'],
-                    'total' => 0.0,
-                    'byComponent' => [],
-                    'byZone' => [],
-                    'byPhase' => ['phase1' => 0.0, 'phase2' => 0.0, 'phase3' => 0.0],
-                ];
-            }
-
-            $componentCode = EnergyPriceComponent::from((int)$row['component_code'])->name;
-            $componentCost = round((float)$row['total_kwh'] * (float)$row['amount'], 6);
-            $phase1Cost = round((float)$row['phase1_kwh'] * (float)$row['amount'], 6);
-            $phase2Cost = round((float)$row['phase2_kwh'] * (float)$row['amount'], 6);
-            $phase3Cost = round((float)$row['phase3_kwh'] * (float)$row['amount'], 6);
-
-            $logs[$key]['costs']['total'] += $componentCost;
-            $logs[$key]['costs']['byComponent'][$componentCode] = ($logs[$key]['costs']['byComponent'][$componentCode] ?? 0) + $componentCost;
-            if ($row['zone_code']) {
-                $logs[$key]['costs']['byZone'][$row['zone_code']] = ($logs[$key]['costs']['byZone'][$row['zone_code']] ?? 0) + $componentCost;
-            }
-            $logs[$key]['costs']['byPhase']['phase1'] += $phase1Cost;
-            $logs[$key]['costs']['byPhase']['phase2'] += $phase2Cost;
-            $logs[$key]['costs']['byPhase']['phase3'] += $phase3Cost;
-        }
-
-        foreach ($logs as &$log) {
-            if ($log['costs']) {
-                $log['costs']['total'] = round($log['costs']['total'], 6);
-                foreach ($log['costs']['byComponent'] as $component => $amount) {
-                    $log['costs']['byComponent'][$component] = round($amount, 6);
-                }
-                foreach ($log['costs']['byZone'] as $zone => $amount) {
-                    $log['costs']['byZone'][$zone] = round($amount, 6);
-                }
-                foreach ($log['costs']['byPhase'] as $phase => $amount) {
-                    $log['costs']['byPhase'][$phase] = round($amount, 6);
-                }
-            }
-        }
-
-        return array_values($logs);
-    }
-
-    private function buildEnergyCostSummaries(int $channelId, array $rows, int $afterTimestamp, int $beforeTimestamp): array {
-        $logs = $this->hydrateCostLogs($rows);
-        $tariffIds = array_values(array_unique(array_filter(array_map(fn(array $log) => $log['tariffId'], $logs))));
-        $pricePeriodIds = array_values(array_unique(array_filter(array_map(fn(array $log) => $log['pricePeriodId'], $logs))));
-        $tariffs = [];
-        foreach ($tariffIds as $tariffId) {
-            $tariff = $this->getMeasurementLogsEntityManager()->find(EnergyTariff::class, $tariffId);
-            if ($tariff) {
-                $tariffs[$tariffId] = $tariff;
-            }
-        }
-        $pricePeriods = [];
-        foreach ($pricePeriodIds as $pricePeriodId) {
-            $pricePeriod = $this->getMeasurementLogsEntityManager()->find(EnergyTariffProfilePricePeriod::class, $pricePeriodId);
-            if ($pricePeriod) {
-                $pricePeriods[$pricePeriodId] = $pricePeriod;
-            }
-        }
-
-        $summaries = [];
-        foreach ($logs as $log) {
-            $context = $this->resolveBillingContext($log, $tariffs, $pricePeriods);
-            $key = $context['key'];
-            if (!isset($summaries[$key])) {
-                $summaries[$key] = [
-                    'periodStart' => $context['periodStart'],
-                    'periodEnd' => $context['periodEnd'],
-                    'timezone' => $context['timezone'],
-                    'usage' => [
-                        'totalKwh' => 0.0,
-                        'byPhase' => ['phase1' => 0.0, 'phase2' => 0.0, 'phase3' => 0.0],
-                    ],
-                    'costs' => [
-                        'currency' => $log['costs']['currency'] ?? 'PLN',
-                        'total' => 0.0,
-                        'byComponent' => [],
-                        'byZone' => [],
-                        'byPhase' => ['phase1' => 0.0, 'phase2' => 0.0, 'phase3' => 0.0],
-                    ],
-                ];
-            }
-
-            $summaries[$key]['usage']['totalKwh'] += $log['usage']['totalKwh'];
-            $summaries[$key]['usage']['byPhase']['phase1'] += $log['usage']['phase1Fae'] / 1000;
-            $summaries[$key]['usage']['byPhase']['phase2'] += $log['usage']['phase2Fae'] / 1000;
-            $summaries[$key]['usage']['byPhase']['phase3'] += $log['usage']['phase3Fae'] / 1000;
-
-            if ($log['costs']) {
-                $summaries[$key]['costs']['total'] += $log['costs']['total'];
-                foreach ($log['costs']['byComponent'] as $component => $amount) {
-                    $summaries[$key]['costs']['byComponent'][$component] = ($summaries[$key]['costs']['byComponent'][$component] ?? 0) + $amount;
-                }
-                foreach ($log['costs']['byZone'] as $zone => $amount) {
-                    $summaries[$key]['costs']['byZone'][$zone] = ($summaries[$key]['costs']['byZone'][$zone] ?? 0) + $amount;
-                }
-                foreach ($log['costs']['byPhase'] as $phase => $amount) {
-                    $summaries[$key]['costs']['byPhase'][$phase] += $amount;
-                }
-            }
-        }
-
-        $this->applyFixedCostsToSummaries($summaries, $channelId, $afterTimestamp, $beforeTimestamp);
-
-        foreach ($summaries as &$summary) {
-            $summary['usage']['totalKwh'] = round($summary['usage']['totalKwh'], 6);
-            foreach ($summary['usage']['byPhase'] as $phase => $amount) {
-                $summary['usage']['byPhase'][$phase] = round($amount, 6);
-            }
-            $summary['costs']['total'] = round($summary['costs']['total'], 6);
-            foreach ($summary['costs']['byComponent'] as $component => $amount) {
-                $summary['costs']['byComponent'][$component] = round($amount, 6);
-            }
-            foreach ($summary['costs']['byZone'] as $zone => $amount) {
-                $summary['costs']['byZone'][$zone] = round($amount, 6);
-            }
-            foreach ($summary['costs']['byPhase'] as $phase => $amount) {
-                $summary['costs']['byPhase'][$phase] = round($amount, 6);
-            }
-        }
-
-        ksort($summaries);
-        return $summaries;
-    }
-
-    private function applyFixedCostsToSummaries(array &$summaries, int $channelId, int $afterTimestamp, int $beforeTimestamp): void {
-        $assignment = $this->findProfileAssignmentForChannel($channelId);
-        if (!$assignment || !$assignment->getProfile()) {
-            return;
-        }
-
-        $rangeStart = new \DateTime('@' . max($afterTimestamp, 0));
-        $rangeStart->setTimezone(new \DateTimeZone('UTC'));
-        $rangeEnd = $beforeTimestamp > 0 ? new \DateTime('@' . $beforeTimestamp) : new \DateTime('now', new \DateTimeZone('UTC'));
-        $rangeEnd->setTimezone(new \DateTimeZone('UTC'));
-
-        foreach ($assignment->getProfile()->getTariffPeriods() as $tariffPeriod) {
-            $tariff = $tariffPeriod->getTariff();
-            if (!$tariff) {
-                continue;
-            }
-            $timezone = new \DateTimeZone($tariff->getConfig()['timezone'] ?? 'UTC');
-            foreach ($tariffPeriod->getPricePeriods() as $pricePeriod) {
-                $pricePeriodStart = $this->maxDateTime(
-                    clone $rangeStart,
-                    $tariffPeriod->getValidFrom(),
-                    $pricePeriod->getValidFrom()
-                );
-                $pricePeriodEnd = $this->minDateTime(
-                    clone $rangeEnd,
-                    $tariffPeriod->getValidTo(),
-                    $pricePeriod->getValidTo()
-                );
-                $start = $pricePeriodStart > $rangeStart ? $pricePeriodStart : clone $rangeStart;
-                $end = $pricePeriodEnd < $rangeEnd ? $pricePeriodEnd : clone $rangeEnd;
-                if ($start >= $end) {
-                    continue;
-                }
-
-                $billingContext = $this->resolveBillingPeriodForTimestamp(
-                    $start->getTimestamp(),
-                    $timezone,
-                    $pricePeriod->getValidFrom(),
-                    $pricePeriod->getBillingPeriodLength(),
-                    $pricePeriod->getBillingPeriodUnit()
-                );
-                while (strtotime($billingContext['periodStart']) < $end->getTimestamp()) {
-                    $summaryKey = $billingContext['periodStart'] . '|' . $billingContext['timezone'];
-                    if (!isset($summaries[$summaryKey])) {
-                        $summaries[$summaryKey] = [
-                            'periodStart' => $billingContext['periodStart'],
-                            'periodEnd' => $billingContext['periodEnd'],
-                            'timezone' => $billingContext['timezone'],
-                            'usage' => ['totalKwh' => 0.0, 'byPhase' => ['phase1' => 0.0, 'phase2' => 0.0, 'phase3' => 0.0]],
-                            'costs' => [
-                                'currency' => $pricePeriod->getCurrency(),
-                                'total' => 0.0,
-                                'byComponent' => [],
-                                'byZone' => [],
-                                'byPhase' => ['phase1' => 0.0, 'phase2' => 0.0, 'phase3' => 0.0],
-                            ],
-                        ];
-                    }
-
-                    foreach ($pricePeriod->getItems() as $item) {
-                        if ($item->getUnit() === EnergyPriceUnit::KWH) {
-                            continue;
-                        }
-                        if ($item->getUnit() === EnergyPriceUnit::PERIOD) {
-                            $this->addSummaryCost($summaries[$summaryKey], $item->getComponentCode(), (float)$item->getAmount());
-                        } else {
-                            $unitCount = $this->countOverlappingFixedUnits(
-                                $billingContext['periodStart'],
-                                $billingContext['periodEnd'],
-                                $start,
-                                $end,
-                                $timezone,
-                                $item->getUnit()
-                            );
-                            $this->addSummaryCost($summaries[$summaryKey], $item->getComponentCode(), (float)$item->getAmount() * $unitCount);
-                        }
-                    }
-
-                    $next = new \DateTime($billingContext['periodEnd'], new \DateTimeZone('UTC'));
-                    $billingContext = $this->resolveBillingPeriodForTimestamp(
-                        $next->getTimestamp(),
-                        $timezone,
-                        $pricePeriod->getValidFrom(),
-                        $pricePeriod->getBillingPeriodLength(),
-                        $pricePeriod->getBillingPeriodUnit()
-                    );
-                }
-            }
-        }
-    }
-
-    private function addSummaryCost(array &$summary, EnergyPriceComponent $componentCode, float $amount): void {
-        $summary['costs']['total'] += $amount;
-        $summary['costs']['byComponent'][$componentCode->name] = ($summary['costs']['byComponent'][$componentCode->name] ?? 0) + $amount;
-    }
-
-    private function countOverlappingFixedUnits(
-        string $periodStart,
-        string $periodEnd,
-        \DateTime $rangeStart,
-        \DateTime $rangeEnd,
-        \DateTimeZone $timezone,
-        EnergyPriceUnit $unit
-    ): int {
-        $periodStartLocal = new \DateTime($periodStart, new \DateTimeZone('UTC'));
-        $periodEndLocal = new \DateTime($periodEnd, new \DateTimeZone('UTC'));
-        $periodStartLocal->setTimezone($timezone);
-        $periodEndLocal->setTimezone($timezone);
-        $overlapStart = clone $periodStartLocal;
-        $overlapEnd = clone $periodEndLocal;
-        $rangeStartLocal = clone $rangeStart;
-        $rangeStartLocal->setTimezone($timezone);
-        $rangeEndLocal = clone $rangeEnd;
-        $rangeEndLocal->setTimezone($timezone);
-        if ($rangeStartLocal > $overlapStart) {
-            $overlapStart = clone $rangeStartLocal;
-        }
-        if ($rangeEndLocal < $overlapEnd) {
-            $overlapEnd = clone $rangeEndLocal;
-        }
-        if ($overlapStart >= $overlapEnd) {
-            return 0;
-        }
-        $cursor = clone $periodStartLocal;
-        $count = 0;
-        while ($cursor < $periodEndLocal) {
-            $next = $this->advanceDateTime(clone $cursor, 1, $this->mapPriceUnitToBillingPeriodUnit($unit));
-            if ($next > $periodEndLocal) {
-                $next = clone $periodEndLocal;
-            }
-            if ($next > $overlapStart && $cursor < $overlapEnd) {
-                $count++;
-            }
-            $cursor = $next;
-        }
-        return $count;
-    }
-
-    private function resolveBillingContext(array $log, array $tariffs, array $pricePeriods): array {
-        $pricePeriod = $log['pricePeriodId'] ? ($pricePeriods[$log['pricePeriodId']] ?? null) : null;
-        $tariff = $log['tariffId'] ? ($tariffs[$log['tariffId']] ?? null) : null;
-        $timezone = new \DateTimeZone($tariff?->getConfig()['timezone'] ?? 'UTC');
-        if (!$pricePeriod) {
-            return $this->resolveDefaultBillingPeriodForTimestamp($log['slotStartTimestamp'], $timezone);
-        }
-        return $this->resolveBillingPeriodForTimestamp(
-            $log['slotStartTimestamp'],
-            $timezone,
-            $pricePeriod->getValidFrom(),
-            $pricePeriod->getBillingPeriodLength(),
-            $pricePeriod->getBillingPeriodUnit()
-        );
-    }
-
-    private function resolveBillingPeriodForTimestamp(
-        int $timestamp,
-        \DateTimeZone $timezone,
-        ?\DateTime $billingAnchorUtc,
-        int $billingPeriodLength,
-        BillingPeriodUnit $billingPeriodUnit
-    ): array {
-        if ($billingAnchorUtc === null) {
-            return $this->resolveNaturalBillingPeriodForTimestamp($timestamp, $timezone, $billingPeriodLength, $billingPeriodUnit);
-        }
-
-        $local = new \DateTime('@' . $timestamp);
-        $local->setTimezone($timezone);
-        $periodStartLocal = clone $billingAnchorUtc;
-        $periodStartLocal->setTimezone($timezone);
-        $periodEndLocal = $this->advanceDateTime(clone $periodStartLocal, $billingPeriodLength, $billingPeriodUnit);
-
-        while ($local < $periodStartLocal) {
-            $periodEndLocal = clone $periodStartLocal;
-            $periodStartLocal = $this->advanceDateTime(clone $periodStartLocal, -$billingPeriodLength, $billingPeriodUnit);
-        }
-
-        while ($local >= $periodEndLocal) {
-            $periodStartLocal = clone $periodEndLocal;
-            $periodEndLocal = $this->advanceDateTime(clone $periodEndLocal, $billingPeriodLength, $billingPeriodUnit);
-        }
-
-        $periodStartUtc = clone $periodStartLocal;
-        $periodStartUtc->setTimezone(new \DateTimeZone('UTC'));
-        $periodEndUtc = clone $periodEndLocal;
-        $periodEndUtc->setTimezone(new \DateTimeZone('UTC'));
-
-        return [
-            'key' => $periodStartUtc->format(\DateTime::ATOM) . '|' . $timezone->getName(),
-            'periodStart' => $periodStartUtc->format(\DateTime::ATOM),
-            'periodEnd' => $periodEndUtc->format(\DateTime::ATOM),
-            'timezone' => $timezone->getName(),
-        ];
-    }
-
-    private function resolveNaturalBillingPeriodForTimestamp(
-        int $timestamp,
-        \DateTimeZone $timezone,
-        int $billingPeriodLength,
-        BillingPeriodUnit $billingPeriodUnit
-    ): array {
-        $local = new \DateTime('@' . $timestamp);
-        $local->setTimezone($timezone);
-        $periodStartLocal = $this->alignNaturalBillingPeriodStart(clone $local, $billingPeriodLength, $billingPeriodUnit);
-        $periodEndLocal = $this->advanceDateTime(clone $periodStartLocal, $billingPeriodLength, $billingPeriodUnit);
-
-        $periodStartUtc = clone $periodStartLocal;
-        $periodStartUtc->setTimezone(new \DateTimeZone('UTC'));
-        $periodEndUtc = clone $periodEndLocal;
-        $periodEndUtc->setTimezone(new \DateTimeZone('UTC'));
-
-        return [
-            'key' => $periodStartUtc->format(\DateTime::ATOM) . '|' . $timezone->getName(),
-            'periodStart' => $periodStartUtc->format(\DateTime::ATOM),
-            'periodEnd' => $periodEndUtc->format(\DateTime::ATOM),
-            'timezone' => $timezone->getName(),
-        ];
-    }
-
-    private function resolveDefaultBillingPeriodForTimestamp(int $timestamp, \DateTimeZone $timezone): array {
-        return $this->resolveNaturalBillingPeriodForTimestamp($timestamp, $timezone, 1, BillingPeriodUnit::MONTH);
-    }
-
-    private function alignNaturalBillingPeriodStart(\DateTime $dateTime, int $length, BillingPeriodUnit $unit): \DateTime {
-        $length = max($length, 1);
-
-        return match ($unit) {
-            BillingPeriodUnit::DAY => $this->alignDayBillingPeriodStart($dateTime, $length),
-            BillingPeriodUnit::WEEK => $this->alignWeekBillingPeriodStart($dateTime, $length),
-            BillingPeriodUnit::MONTH => $this->alignMonthBillingPeriodStart($dateTime, $length),
-            BillingPeriodUnit::YEAR => $this->alignYearBillingPeriodStart($dateTime, $length),
-        };
-    }
-
-    private function alignDayBillingPeriodStart(\DateTime $dateTime, int $length): \DateTime {
-        $dateTime->setTime(0, 0, 0);
-        if ($length === 1) {
-            return $dateTime;
-        }
-
-        $epoch = new \DateTime('1970-01-01 00:00:00', $dateTime->getTimezone());
-        $dayOffset = (int)$epoch->diff($dateTime)->format('%r%a');
-        $remainder = (($dayOffset % $length) + $length) % $length;
-        return $remainder ? $dateTime->modify(sprintf('-%d day', $remainder)) : $dateTime;
-    }
-
-    private function alignWeekBillingPeriodStart(\DateTime $dateTime, int $length): \DateTime {
-        $dateTime->setTime(0, 0, 0);
-        $dateTime->modify('monday this week');
-        if ($length === 1) {
-            return $dateTime;
-        }
-
-        $epoch = new \DateTime('1970-01-05 00:00:00', $dateTime->getTimezone());
-        $dayOffset = (int)$epoch->diff($dateTime)->format('%r%a');
-        $weekOffset = intdiv($dayOffset, 7);
-        $remainder = (($weekOffset % $length) + $length) % $length;
-        return $remainder ? $dateTime->modify(sprintf('-%d week', $remainder)) : $dateTime;
-    }
-
-    private function alignMonthBillingPeriodStart(\DateTime $dateTime, int $length): \DateTime {
-        $dateTime->setDate((int)$dateTime->format('Y'), (int)$dateTime->format('m'), 1);
-        $dateTime->setTime(0, 0, 0);
-        if ($length === 1) {
-            return $dateTime;
-        }
-
-        $monthOffset = ((int)$dateTime->format('Y') * 12) + ((int)$dateTime->format('n') - 1);
-        $remainder = (($monthOffset % $length) + $length) % $length;
-        return $remainder ? $dateTime->modify(sprintf('-%d month', $remainder)) : $dateTime;
-    }
-
-    private function alignYearBillingPeriodStart(\DateTime $dateTime, int $length): \DateTime {
-        $dateTime->setDate((int)$dateTime->format('Y'), 1, 1);
-        $dateTime->setTime(0, 0, 0);
-        if ($length === 1) {
-            return $dateTime;
-        }
-
-        $year = (int)$dateTime->format('Y');
-        $remainder = (($year % $length) + $length) % $length;
-        return $remainder ? $dateTime->modify(sprintf('-%d year', $remainder)) : $dateTime;
-    }
-
-    private function advanceDateTime(\DateTime $dateTime, int $length, BillingPeriodUnit $unit): \DateTime {
-        $sign = $length >= 0 ? '+' : '-';
-        $absoluteLength = abs($length);
-        return match ($unit) {
-            BillingPeriodUnit::DAY => $dateTime->modify(sprintf('%s%d day', $sign, $absoluteLength)),
-            BillingPeriodUnit::WEEK => $dateTime->modify(sprintf('%s%d week', $sign, $absoluteLength)),
-            BillingPeriodUnit::MONTH => $dateTime->modify(sprintf('%s%d month', $sign, $absoluteLength)),
-            BillingPeriodUnit::YEAR => $dateTime->modify(sprintf('%s%d year', $sign, $absoluteLength)),
-        };
-    }
-
-    private function mapPriceUnitToBillingPeriodUnit(EnergyPriceUnit $unit): BillingPeriodUnit {
-        return match ($unit) {
-            EnergyPriceUnit::DAY => BillingPeriodUnit::DAY,
-            EnergyPriceUnit::WEEK => BillingPeriodUnit::WEEK,
-            EnergyPriceUnit::MONTH => BillingPeriodUnit::MONTH,
-            default => throw new \InvalidArgumentException('Unsupported fixed price unit.'),
-        };
+        return $this->energyCostRowFetcher->fetchCostRows($channelId, $afterTimestamp, $beforeTimestamp, $orderDesc, $limit, $offset);
     }
 
     private function ensureApiVersion24(Request $request): void {
